0.4.1
-Yet another complete refactor of requests.py.  Feeling good about its
speed, though.
-Changed app.yaml version number... poking blindly at implementing
versioning.
