+++ app.yaml
@@ -1,5 +1,5 @@
 application: wclstats
-version: 1
+version: 0-4
 runtime: python27
 api_version: 1
 threadsafe: yes
